認証ロジックをlib/auth.tsに移動し、route.tsを簡素化。ログインエラーハンドリングを改善。
